Remove JavaLayerConfigurations
